Edited values to make losing a war more damaging to your army, lessening the effect(s) of lucky fighting sprees.
Essentially, this means that if you fight a lot of wars in the aim of winning one, or win a lot and then lose one, you get punished more. This way, you actually have to explore, conscript, buy, or hold a parade in order to be able to have a good shot at invading more countries.